LayoutTests/imported/w3c/web-platform-tests/html/dom/render-blocking/remove-attr-style-keeps-blocking.tentative-expected.txt

Some text

PASS Rendering is blocked before render-blocking resources are loaded
PASS Render-blocking stylesheet is applied
